token generation, templates and alerts

+++ frontend/views.py
@@ -2,6 +2,7 @@ from django.contrib.auth import get_user_model
 from django.db import transaction
 from django.shortcuts import render, get_object_or_404
 from django.views.generic import View
+from control import Status
 from frontend.forms import EditProfileForm
 
 BotUser = get_user_model()
@@ -11,20 +12,15 @@ class ProfileOverView(View):
     template = "frontend/profile_overview.html"
 
     def get(self, request, *args, **kwargs):
-        user = get_object_or_404(BotUser, username=request.user.username)
-        self.data = {
-            "user": user,
-            "message": "something to tell you whatever that might be",
-            "alert": True,
-        }
-        return render(request, self.template, self.data)
+        data = Status()
+        return render(request, self.template, data)
 
 
 class EditProfile(View):
     template = "frontend/edit_profile.html"
 
     def get(self, request):
-        user = get_object_or_404(BotUser, id=request.user.id)
+        user = get_object_or_404(BotUser, username=request.user.username)
         initial = {
             "first_name": user.first_name or 'To be set',
             "last_name": user.last_name or 'To be set',
@@ -33,16 +29,13 @@ class EditProfile(View):
             "host": user.host,
             "about": user.about,
         }
-        form = EditProfileForm(initial=initial)
-        data = { "form": form, "user": user, "success": True, "alert": "success", "message": "Something"}
-        return render(request, self.template, data)
-
+        return render(request, self.template, Status(form=EditProfileForm(initial=initial)))
 
     @transaction.atomic()
     def post(self, request):
         form = EditProfileForm(request.POST)
         user = get_object_or_404(BotUser, id=request.user.id)
-
+        data = Status(form=form, alert=True, success=False, message='Error updating profile!')
         if form.is_valid():
             user.first_name = form.clean()['first_name']
             user.last_name = form.clean()['last_name']
@@ -51,21 +44,53 @@ class EditProfile(View):
             user.host = form.clean()['host']
             user.about = form.clean()['about']
             user.save(force_update=True)
-            return render(request, "frontend/profile_overview.html", {"user": user})
-        return render(request, self.template, {"user": user})
+
+            data.update({'form': form, 'success': True, 'alert': True, 'message': 'User profile updated!'})
+            return render(request, "frontend/profile_overview.html", data)
+        return render(request, self.template, data)
 
 
 class DeleteProfile(View):
     template = "frontend/delete_profile.html"
 
     def get(self, request):
-        user = get_object_or_404(BotUser, id=request.user.id)
-        return render(request, self.template, {"user": user})
+        return render(request, self.template, Status())
 
     @transaction.atomic()
     def post(self, request):
         user = get_object_or_404(BotUser, id=request.user.id)
-        if request.POST.get("delete", None) == "1":
+        data = Status(user=user, alert=True, success=False, message='Error updating profile!')
+
+        if request.POST.get("delete", None) == "0":
+            data.update({"message": "Deletion canceled!", "alert": True, "success": False})
+            return render(request, "frontend/profile_overview.html", data)
+
+        elif request.POST.get("delete", None) == "1":
+            if user.is_superuser:
+                data.update({
+                    "message": "Superusers can't delete themselves in the profile editor!",
+                    "alert": True,
+                    "success": False})
+                return render(request, self.template, data)
+
             user.delete()
-            return render(request, "auth/login.html")
-        return render(request, self.template, {"user": user})
+            data.update({'success': True, 'alert': True, 'message': 'User profile deleted!'})
+            return render(request, "auth/login.html", data)
+        return render(request, self.template, data)
+
+
+class RegenerateToken(View):
+    template = "frontend/profile_overview.html"
+
+    def get(self, request):
+        return render(request, self.template, Status())
+
+    def post(self, request):
+        user = get_object_or_404(BotUser, username=request.user.username)
+        data = Status(user=user, alert=True, success=False, message='Error regenerating token!')
+
+        if request.POST.get("regenerate_token", None) == "1":
+            user.registration_token = None
+            user.save()
+            data.update({'success': True, 'alert': True, 'message': 'Registration token regenerated!'})
+        return render(request, self.template, data)
